requirements.txt
pyfuse3==3.1.1
trio==0.17.0
defusedxml==0.6.0
pycryptodome==3.9.9